JobOutput: extract multiple helper functions
